improve the robustness of the memory reading
Sometimes, the offsets for some fields are slightly wrong (usually by
one byte). This commit tries to get the fields after constructing the
ManagedClass instance. If it fails, it re-reads the memory to get the
new, correct value of the memory pointer.
The drawnback is twofolds:
- it takes longer to instantiate the class
- it removes the benefit of lazy fields

Nevertheless, the fact that this removes the need for module-wide
resets is a huge benefit.